Auto PR: Regenerating the Go SDK (64badb74c5ba8b8ace25aa16448213a638e6e478)

+++ resource-manager/web/2023-12-01/appserviceenvironments/model_functionsscaleandconcurrency.go
@@ -5,7 +5,7 @@ package appserviceenvironments
 
 type FunctionsScaleAndConcurrency struct {
 	AlwaysReady          *[]FunctionsAlwaysReadyConfig         `json:"alwaysReady,omitempty"`
-	InstanceMemoryMB     *float64                              `json:"instanceMemoryMB,omitempty"`
-	MaximumInstanceCount *float64                              `json:"maximumInstanceCount,omitempty"`
+	InstanceMemoryMB     *int64                                `json:"instanceMemoryMB,omitempty"`
+	MaximumInstanceCount *int64                                `json:"maximumInstanceCount,omitempty"`
 	Triggers             *FunctionsScaleAndConcurrencyTriggers `json:"triggers,omitempty"`
 }
